Fixing issues with overflow on generating int, short, etc. when the range is T.Min to T.Max. Added the ability to generate generic arrays, lists, IEnumerable, and dictionaries. Cleaned up Builder's GetGenerator methods. Added a method to get an IEnumerable from Random based on the Type.

+++ src/Mirage/Generators/Default/IntGeneratorAttribute.cs
@@ -70,7 +70,7 @@ namespace Mirage.Generators
         /// <returns>A randomly generated object of the specified type</returns>
         public int Next(Random rand, int min, int max)
         {
-            return min + (int)((max - min) * rand.NextDouble());
+            return (int)(min + (long)(((long)max - (long)min) * rand.NextDouble()));
         }
 
         /// <summary>
